Break down the AST processing code
