+++ packages/kbn-interpreter/src/common/lib/ast/from_expression.test.js
@@ -7,9 +7,9 @@
  */
 
 import { fromExpression } from '@kbn/interpreter';
-import { getType } from './get_type';
+import { getType } from '../get_type';
 
-describe('ast fromExpression', () => {
+describe('fromExpression', () => {
   describe('invalid expression', () => {
     it('throws with invalid expression', () => {
       const check = () => fromExpression('wat!');
