+++ packages/kbn-interpreter/src/common/lib/ast/safe_element_from_expression.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { fromExpression } from './from_expression';
+
+// TODO: OMG This is so bad, we need to talk about the right way to handle bad expressions since some are element based and others not
+export function safeElementFromExpression(expression: string) {
+  try {
+    return fromExpression(expression);
+  } catch (e) {
+    return fromExpression(
+      `markdown
+"## Crud.
+Canvas could not parse this element's expression. I am so sorry this error isn't more useful. I promise it will be soon.
+
+Thanks for understanding,
+#### Management
+"`
+    );
+  }
+}
